Удалена функция инициализации светодиода. Удален define для управления светодиодом.

+++ Software/stm8s_it.c
@@ -457,8 +457,7 @@ INTERRUPT_HANDLER(TIM6_UPD_OVF_TRG_IRQHandler, 23)
   /* In order to detect unexpected events during development,
      it is recommended to set a breakpoint on the following instruction.
   */
-   
-   extern int32_t msVar;
+   extern uint8_t msVar;
    
    TIM4_ClearITPendingBit(TIM4_IT_UPDATE);
